move fragment classes to ui package
