+++ app/app/src/main/java/com/ar/enbaldeapp/ui/NotiificationsFragment.java
@@ -1,19 +1,20 @@
-package com.ar.enbaldeapp;
+package com.ar.enbaldeapp.ui;
 
 import android.os.Bundle;
-
 import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.ar.enbaldeapp.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link orders#newInstance} factory method to
+ * Use the {@link NotiificationsFragment#newInstance} factory method to
  * create an instance of this fragment.
+ *
  */
-public class orders extends Fragment {
+public class NotiificationsFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,26 +25,26 @@ public class orders extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public orders() {
-        // Required empty public constructor
-    }
-
     /**
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment orders.
+     * @return A new instance of fragment NotiificationsFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static orders newInstance(String param1, String param2) {
-        orders fragment = new orders();
+    public static NotiificationsFragment newInstance(String param1, String param2) {
+        NotiificationsFragment fragment = new NotiificationsFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
         fragment.setArguments(args);
         return fragment;
+    }
+
+    public NotiificationsFragment() {
+        // Required empty public constructor
     }
 
     @Override
@@ -59,6 +60,6 @@ public class orders extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_orders, container, false);
+        return inflater.inflate(R.layout.fragment_notiifications, container, false);
     }
 }
