+++ app/app/src/main/res/layout/activity_pay.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Pay">
+    tools:context=".ui.Pay">
 <!--
     <TextView
         android:id="@+id/textView6"
