+++ app/app/src/main/res/layout/fragment_notiifications.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NotiificationsFragment">
+    tools:context=".ui.NotiificationsFragment">
 
     <TextView
         android:id="@+id/textView4"
